Add actions for deleting/restoring nodes
